Changed reading to allow 'I-X' to follow other tags than just 'B-X', by interpreting them as 'B' tags. This means that the output of machine learners can be loaded, rather than just the gold standard.
Changed reader to return the top level tree from the stack, not the bottom,
which often leaves out parts of unfinished sentences. The gold standard has a
number of erroneous newlines, which caused parts of the sentences to
disappear, as they're not terminated with a full stop.


svn/trunk/nltk@2142