First pass at CSV data file for NLTK corpora.
Fields:
* nltk_directory
* short_name
* long_name
* contents
* derived_from
* reader
* licenser1
* licenser2
* see_also
* contact_name
* contact_email

The short name should be the canonical term we use to refer to the
corpus within The Book. At the moment, the default short name is the
same as the directory name in the distribution -- needs more consideration.

The 'contents' field should perhaps be thought about more; a full
listing is impractical in many cases (e.g. Timit), but it maybe useful
to indicate if only specific subparts of the full corpus are distributed.

List items within fields are separated by '::'

The 'see_also' field should contain a list of URLs.


svn/trunk@4552