Fixed some deficiencies in the logic syntax:     *LambdaExpression now holds exactly one bound variable, instead of a list of bound variables     *'\x.\y.E' is parsed exactly like '\x y.E'. So now, '\x.\y.R(x,y)(a)' == '(\x.\y.R(x,y))(a)'     *Functions can be applied to more than one argument in a row. So now, '\x.\y.R(x,y)(a)(b)' == '((\x.\y.R(x,y))(a))(b)'     *Functions can still be applied to tuples, but now the tuples are broken up.  So now, '\x.\y.R(x,y)(a,b)' == '((\x.\y.R(x,y))(a))(b)'
Changed LogicParser to clear the buffer and reset the currentIndex on parse().  This fixes a bug where, if the parser threw an exception, then the parser would still be pointing to the previous parse's data when parse() was called with new data.  This led to problem when one tried to use the same LogicParser object to parse more than one string.

Also cleaned up the unit tests a bit.

svn/branches/first_order_predicate_calc@6082