Updated semantics of variables() and free().
variables() now returns a set of all the variables that are available to be replaced.  This includes free (non-bound) variables as well as predicates.

free() now takes an optional argument, 'indvar_only', that tells the method to return only individual variables.  This argument is True by default.  The method has also been modified to prevent the inclusion of predicates.

svn/trunk@6407