Docbook xlinks are rendered as "Figure X", "Example X", "Section X", "Table X", "Chapter X" etc.  When the underlying text includes the object type, the resulting output has "Section Section X".  Therefore, all instances of these words before a citation in our source text, e.g. "Section sec-foo_", have now been deleted.
A consequence of this is that our HTML and LaTeX output would lack the extra word having the object type.  So rst.py is modified to add this in (just for HTML and LaTeX).

Note that references to linguistic examples (53) and to code callouts, do not get the extra word in any output format.  Occasionally there is prose of the form: "In example (53)", instead of "In (53)".  This is passed through, but a warning message is generated.

SIGH


svn/trunk@7720