Add set_proxy function to nltk.util to allow easy configuration of HTTP proxy.
This function sets up the required handlers for accessing HTTP content through 
a proxy, potentially with authentication. For example:

>>> import nltk
>>> nltk.set_proxy('http://proxy.example.com:3128', ('USERNAME', 'PASSWORD'))
>>> nltk.download()

The username and password fields are optional to support non-authenticating 
proxies. If proxy is set to None then this function will attempt to detect the 
system proxy or raise a ValueError if a proxy can't be discovered.


svn/trunk@8676