Updated docs, doctests, error if text is non-ASCII
The BllipParser class docs now includes references for the parser.

The wrapper will throw an error if given non-ASCII text (at least until
we settle on a way to make it safe).

Added doctests (based on demo()), doctests are skipped if bllipparser
isn't installed.

demo() now has a stub to use nltk.download().  Since nltk_data
doesn't currently have a BLLIP model in it, I still included
bllipparser.ModelFetcher as a fallback in demo() and the doctests
so they'll still work.  We can drop this once nltk_data is updated.