Fix dict-order issue with probability.doctest
The FreqDist class is a subclass of Counter, and while Counter sorts elements by count, "elements with equal counts are ordered arbitrarily" (https://docs.python.org/3/library/collections.html#collections.Counter.most_common).

This arbitrary ordering causes issues with doctests, because doctests require strict string equivalence, and the ordering can vary across Python versions and platforms (i.e., the test may pass for one version of Python on a certain platform, but it is not guaranteed to do so elsewhere).

I decided to fix the doctest by changing the distributions so that no two elements had an equal count for any of the operations. There are other ways to fix the test (e.g., manually sorting the FreqDist by count and then by element, or by comparing the result with expected values and looking for True values), this change still follows the spirit of the original doctest in showcasing the behavior of the class.