Added test cases for remove_handle
Test cases for the following change:
https://github.com/nltk/nltk/pull/1521

Writing the expected results was a little difficult as the behavior around adjacent punctuation and especially "@"s that are not part of Twitter handles is confusing. I am pretty sure the test cases are correct, but if one of them fails when it shouldn't I would be happy to take another look.