Allow iterator in perceptron tagger training
We'd like to allow ``sentences`` to be either a list or an iterator,
the latter being especially important for a large training dataset.
We first initialize an empty list of ``self._sentences``.
Because ``self._make_tagdict(sentences)`` runs regardless, we make
it populate ``self._sentences`` with all the sentences.
This saves the overheard of just iterating through ``sentences`` to
get the list by ``sentences = list(sentences)``.