Addresses #1667; add gleu_score.corpus_gleu()
From an email thread on nltk-dev, we got some more information about
the corpus-level GLEU calculation. I used that to create a
corpus_gleu() function in nltk.translate.gleu_score. I did not change
the functions to accept multiple references per hypothesis because
that would change the API, and it doesn't appear that the authors of
GLEU implemented it (though they suggested it could just be the max
score among the references). If we want to allow it, we could do
something like:

 1. Inspect if the reference argument is `list(str)` or
    `list(list(str))` automatically, and silently adapt to the input.
 2. Allow an optional `multiple_reference` parameter which, when True,
    expects the reference argument to be multiple references.
 3. Wait for the next major version of NLTK and change the API.