changes according to dimazest's advice
I have read the code of find_jar_iter in the internals.py.
If path_to_slf4j is set to 'slf4j-api.jar' and input to find_jar()
directly, the searching process will execute:

    if path_to_jar is not None:
        if os.path.isfile(path_to_jar):
            yielded = True
            yield path_to_jar
        else:
            raise LookupError('Could not find %s jar file at %s' %
                            (name_pattern, path_to_jar))

and ignore the environment variables.  So in the default situation we
need to set path_to_jar to None.