Avoid mention of source and target language in AlignedSent
This keeps AlignedSent general. Also, IBM models use mots as the source
and words as the target, contradicting the documentation of AlignedSent.

Additionally, fix wrong alignment used in the __init__ doctest so that
the order of alignment is (words_index, mots_index)

Credit to DavidNemeskey for identifying this long-standing discrepancy.

Fixes #1888