Evaluate type before comparing equality of WordNet and other objects
When comparing the equality of WordNet objects with other objects, it 
is assumed that the other object is a WordNet object and hence the name 
attribute of the other object is compared with the name attribute of 
the WordNet object even though the other object might not be a WordNet 
object.

This will result in an AttributeError exception to be thrown in the 
case where the other object is not a WordNet object. For instance when 
comparing a WordNet object with a string, since the string does not 
contain a name attribute, an exception is thrown.

To prevent this from happening, the type of the objects should be 
evaluated before checking for equality.