Fix UnboundLocalError bug and add tests for TextTilingTokenizer
In TextTilingTokenizer, when illegal arguments are passed for `similarity_method` or
`smoothing_method`, an `UnboundLocalError` exception is thrown.  This is caused by
local variables being set in an `if` statement with no `else` branch.  This fix adds
an `else` branch that raises an exception if an illegal argument is used.

Additionally, doctests have been added for TextTilingTokenizer.