Fix ConditionalFreqDist mutation on tabulate/plot call
Since self[c] adds c to the ConditionalFreqDist regardless of
whether c is a valid condition or not, calling tabulate or plot
with condtiions not in the cfd puts them in conditions, causing
errors when it's called again. Checking to see whether c is in
self before calling self[c] prevents this problem.