FreqDist ordering based on partial order
Change the implementation of <= and >= for FreqDist
so they are both a partial order.

With the old implementation for a pair of two frequency
distributions each could be bigger than the other.

>>> a = FreqDist('aab')
>>> b = FreqDist('abb')
>>> a > b
True
>>> b > a
True

This is very unintuitive as it ignores both common sense
and the mathematical definition of order.

It should be noted that the new ordering scheme is a not
a total order and that both a <= b and b <= a are False.