Fix padding around commas and colons
The tokenizer checks for non-digits because we don’t want to add padding inside numbers and time formats (e.g., “1,000” or “9:30”). However, the detokenizer’s input will never have a colon/comma with spaces around if it was part of a number/time string. Thus, at this last point we just need to remove left padding in all cases, instead of checking for non-digit or whitespace (to handle cases where there’s a comma followed by double-quotes).

Signed-off-by: Or Sharir <or@sharir.org>