Re-implemented conversion of CFG to Chomsky Normal Form
Conversion to Chomsky Normal Form (CNF), implemented as the
'.chomsky_normal_form()' method in grammar.CFG instances, was
incomplete. In particular, it could not deal with empty productions
or mixed productions (with terminals and non-terminals co-occuring
on the right-hand side). The new implementation removes those
shortcomings. Like in the earlier version, the individual steps of
conversion are carried out by calling class methods.