Changed CFG class to allow for empty list of productions
Context-free grammars can have an empty set of productions.
Conversion to Chomsky Normal Form leads to an empty list of
productions if the language of the CFG is empty (i.e., does not
even contain the empty string - i.e., [S -> S]). The
grammar.CFG class did not allow for this case because some of the
methods to determine properties of the grammar (e.g., self.is_emtpy,
self.is_binary) referred to the minimum or maximum length of the
productions. If the list of productions is emtpy, Python's
'min' and 'max' operators throw an exception. This is fixed now:
the methods determining properties of the grammar are rewritten
using 'all', which succeeds on an empty list. The attributes
self._min_len and self._max_len were left in place and are set to
'None' if there are no productions.