Make sure dispersion plot spans full text length
By default, Matplotlib autoscales the x-axis based on where hits are
found. This may result in an unintuitive visualization of the dispersion
which "zooms in" on the portion of the text with hits, failing to
represent the preceding and/or following stretches with no hits. A more
intuitive visualization is one where the x-axis always spans the full
length of the text.

The most straightforward way to achieve this is via `ax.set_xlim(0,
len(text) - 1)`, but since this disables autoscaling, it unfortunately
also gets rid of the margins added during autoscaling, which add
aesthetically pleasing breathing room to the plot. So instead, we can
explicitly set the data limits via `ax.dataLim`, and re-apply
autoscaling.